+++ web/src/engine/websites/MangaUpGlobal_e2e.ts
@@ -2,17 +2,18 @@
 
 const config = {
     plugin: {
-        id: 'mangaup',
-        title: 'MangaUp'
+        id: 'mangaupglobal',
+        title: 'MangaUp (Global)'
     },
     container: {
         url: 'https://global.manga-up.com/manga/86',
         id: '86',
-        title: 'A Dating Sim of Life or Death'
+        title: 'A Dating Sim of Life or Death',
+        timeout: 10000
     },
     child: {
         id: '6285',
-        title: 'Chapter 1.1'
+        title: 'Chapter 1.1',
     },
     entry: {
         index: 0,
